修复获取不到学期和已评价课程有重复的BUG
--HG--
branch : develement

+++ evaluation-service/commons/src/main/java/net/tusdasa/evaluation/entity/StudentEvaluation.java
@@ -42,7 +42,7 @@ public class StudentEvaluation implements Serializable {
     /**
      * 当前课程Id
      */
-    private Integer courseId;
+    private Long courseId;
 
     /**
      * 总分 由分析微服务计算
